# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Радар вузов API
  description: '

    CRM контроля взаимодействия ИТ Школы Ростелекома с вузами.


    Все методы `/api/v1` требуют `Authorization: Bearer <access_token>` из Keycloak.

    Ошибки возвращаются в формате `application/problem+json` с полем `code`

    из каталога в ARCHITECTURE.md.

    '
  version: 0.1.0
paths:
...
  /api/v1/workflows/default:
    get:
      tags:
      - workflow
      summary: Опубликованная версия базового workflow
      operationId: read_default_workflow_api_v1_workflows_default_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - HTTPBearer: []
  /api/v1/workflows/default/norms:
    get:
      tags:
      - workflow
      summary: Нормы этапов с подсказками по истории
      operationId: read_norms_api_v1_workflows_default_norms_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/StageNormOut'
                type: array
                title: Response Read Norms Api V1 Workflows Default Norms Get
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - HTTPBearer: []
  /api/v1/workflows/default/norms/{stage_code}:
    put:
      tags:
      - workflow
      summary: Задать норму этапа вручную
      operationId: put_norm_api_v1_workflows_default_norms__stage_code__put
      security:
      - HTTPBearer: []
      parameters:
      - name: stage_code
        in: path
        required: true
        schema:
          type: string
          title: Stage Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NormUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StageNormOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/workflows/default/norms/{stage_code}/accept-suggestion:
    post:
      tags:
      - workflow
      summary: Принять подсказку нормы
      description: Нормой становится 80-й перцентиль завершённых этапов.
      operationId: post_accept_suggestion_api_v1_workflows_default_norms__stage_code__accept_suggestion_post
      security:
      - HTTPBearer: []
      parameters:
      - name: stage_code
        in: path
        required: true
        schema:
          type: string
          title: Stage Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StageNormOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/workflows:
    get:
      tags:
      - workflow
      summary: Процессы и группы, которые по ним работают
      description: У каждой группы контрагентов свой процесс; черновик изменений виден, если начат.
      operationId: read_workflows_api_v1_workflows_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WorkflowSummaryOut'
                type: array
                title: Response Read Workflows Api V1 Workflows Get
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - HTTPBearer: []
    post:
      tags:
      - workflow
      summary: Создать шаблон процесса
      operationId: post_template_api_v1_workflows_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowTemplateCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowTemplateOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - HTTPBearer: []
  /api/v1/workflows/{template_id}:
    get:
      tags:
      - workflow
      summary: Действующая схема процесса
      operationId: read_workflow_api_v1_workflows__template_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Template Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/workflows/{template_id}/norms:
    get:
      tags:
      - workflow
      summary: Нормы этапов процесса с подсказками по истории
      operationId: read_template_norms_api_v1_workflows__template_id__norms_get
      security:
      - HTTPBearer: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Template Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StageNormOut'
                title: Response Read Template Norms Api V1 Workflows  Template Id  Norms Get
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/workflows/{template_id}/norms/{stage_code}:
    put:
      tags:
      - workflow
      summary: Задать норму этапа процесса вручную
      operationId: put_template_norm_api_v1_workflows__template_id__norms__stage_code__put
      security:
      - HTTPBearer: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Template Id
      - name: stage_code
        in: path
        required: true
        schema:
          type: string
          title: Stage Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NormUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StageNormOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/workflows/{template_id}/norms/{stage_code}/accept-suggestion:
    post:
      tags:
      - workflow
      summary: Принять подсказку нормы этапа процесса
      operationId: post_template_accept_suggestion_api_v1_workflows__template_id__norms__stage_code__accept_suggestion_post
      security:
      - HTTPBearer: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Template Id
      - name: stage_code
        in: path
        required: true
        schema:
          type: string
          title: Stage Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StageNormOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/workflows/{template_id}/versions:
    post:
      tags:
      - workflow
      summary: Создать черновик версии
      description: 'Черновик копирует последнюю версию шаблона: править проще, чем собирать заново.'
      operationId: post_version_api_v1_workflows__template_id__versions_post
      security:
      - HTTPBearer: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Template Id
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/workflow-versions/{version_id}:
    patch:
      tags:
      - workflow
      summary: Изменить черновик версии
      description: Этапы, их порядок, нормы, требования к документам и правила переходов.
      operationId: patch_workflow_version_api_v1_workflow_versions__version_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Version Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: WF_VERSION_NOT_DRAFT
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/stages/{stage_id}:
    patch:
      tags:
      - workflow
      summary: Переименовать этап
      description: Единственное изменение, разрешённое в действующей схеме. Переименование статуса — чувствительная
        операция, поэтому доступно только администратору.
      operationId: patch_stage_api_v1_stages__stage_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: stage_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Stage Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StageRename'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StageRef'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/workflow-versions/{version_id}/publish-preview:
    post:
      tags:
      - workflow
      summary: Предпросмотр публикации
      description: Ничего не меняет. Показывает переименованные и добавленные этапы, удалённые этапы с
        числом открытых записей и этапом, куда они переедут, и нужен ли администратор. Данные для окна
        подтверждения перед публикацией.
      operationId: post_publish_preview_api_v1_workflow_versions__version_id__publish_preview_post
      security:
      - HTTPBearer: []
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Version Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishPreview'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR · WF_MIGRATION_MAP_INCOMPLETE
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: WF_VERSION_NOT_DRAFT
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/workflow-versions/{version_id}/publish:
    post:
      tags:
      - workflow
      summary: Опубликовать изменения процесса
      description: Все открытые взаимодействия переезжают на новую схему переходом `migration`, прежняя
        схема становится `retired`. Записи с удалённого этапа переходят на ближайший предыдущий этап,
        а если его нет — на следующий; `migration_map` задаёт другой этап явно. Черновик с переименованием
        этапов публикует только администратор.
      operationId: post_publish_api_v1_workflow_versions__version_id__publish_post
      security:
      - HTTPBearer: []
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Version Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR · WF_MIGRATION_MAP_INCOMPLETE
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: WF_VERSION_NOT_DRAFT
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
...
    NormUpdate:
      properties:
        norm_days:
          type: integer
          maximum: 365.0
          minimum: 1.0
          title: Norm Days
          description: Новая норма этапа в днях
      type: object
      required:
      - norm_days
      title: NormUpdate
...
    Problem:
      properties:
        type:
          type: string
          title: Type
          default: about:blank
        title:
          type: string
          title: Title
        status:
          type: integer
          title: Status
        code:
          $ref: '#/components/schemas/ErrorCode'
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
        trace_id:
          type: string
          title: Trace Id
        errors:
          items:
            $ref: '#/components/schemas/FieldError'
          type: array
          title: Errors
          default: []
      type: object
      required:
      - title
      - status
      - code
      - trace_id
      title: Problem
      description: Тело ответа с ошибкой.
...
    StageNormOut:
      properties:
        stage_code:
          type: string
          title: Stage Code
        stage_name:
          type: string
          title: Stage Name
        norm_days:
          type: integer
          title: Norm Days
        source:
          type: string
          title: Source
          description: manual — задана вручную, suggested — принята подсказка
        suggested_median_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Suggested Median Days
          description: Типичный срок этапа по завершённым переходам
        suggested_percentile_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Suggested Percentile Days
          description: '80-й перцентиль: срок, в который укладывается большинство'
        sample_size:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sample Size
          description: Сколько завершённых этапов легло в подсказку
      type: object
      required:
      - stage_code
      - stage_name
      - norm_days
      - source
      - suggested_median_days
      - suggested_percentile_days
      - sample_size
      title: StageNormOut
...
    StageRename:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 2
          pattern: ^[^\x00-\x1f\x7f]+$
          title: Name
      type: object
      required:
      - name
      title: StageRename
...
    WorkflowOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        template_id:
          type: string
          format: uuid
          title: Template Id
        name:
          type: string
          title: Name
        version_no:
          type: integer
          title: Version No
        stages:
          items:
            $ref: '#/components/schemas/StageOut'
          type: array
          title: Stages
        transitions:
          items:
            $ref: '#/components/schemas/TransitionRuleOut'
          type: array
          title: Transitions
      type: object
      required:
      - id
      - template_id
      - name
      - version_no
      - stages
      - transitions
      title: WorkflowOut
    WorkflowSummaryOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        is_default:
          type: boolean
          title: Is Default
        published_version_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Published Version Id
        version_no:
          anyOf:
          - type: integer
          - type: 'null'
          title: Version No
        draft_version_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Draft Version Id
          description: Черновик изменений, если начат
        groups:
          items:
            $ref: '#/components/schemas/GroupRef'
          type: array
          title: Groups
      type: object
      required:
      - id
      - name
      - is_default
      - published_version_id
      - version_no
      - draft_version_id
      - groups
      title: WorkflowSummaryOut
      description: 'Шаблон процесса: действующая схема, черновик изменений и группы, которые по нему работают.'
    WorkflowTemplateCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 3
          title: Name
      type: object
      required:
      - name
      title: WorkflowTemplateCreate
...
  securitySchemes:
    HTTPBearer:
      type: http
      description: Access token из Keycloak
      scheme: bearer